fix/service 추상화
